Move uix registration file to the uix package
